+++ phpunit/tests/fonts/font-library/wpRestFontLibraryController/uninstallFonts.php
@@ -70,7 +70,6 @@ class Tests_Fonts_WPRESTFontLibraryController_UninstallFonts extends WP_REST_Fon
 		$uninstall_request = new WP_REST_Request( 'DELETE', '/wp/v2/fonts' );
 		$uninstall_request->set_param( 'fontFamilies', $font_families_to_uninstall );
 		$response = rest_get_server()->dispatch( $uninstall_request );
-		echo ( print_r( $response->get_data(), true ) );
 		$this->assertSame( 200, $response->get_status(), 'The response status is not 200.' );
 	}
 
@@ -91,7 +90,7 @@ class Tests_Fonts_WPRESTFontLibraryController_UninstallFonts extends WP_REST_Fon
 
 		$uninstall_request->set_param( 'fontFamilies', $non_existing_font_data );
 		$response = rest_get_server()->dispatch( $uninstall_request );
-		$response->get_data();
-		$this->assertSame( 500, $response->get_status(), 'The response status is not 500.' );
+		$data     = $response->get_data();
+		$this->assertCount( 2, $data['errors'], 'The response should have 2 errors, one for each font family uninstall failure.' );
 	}
 }