+++ test/native/setup.js
@@ -129,6 +129,10 @@ jest.mock( '@wordpress/react-native-bridge', () => {
 		generateHapticFeedback: jest.fn(),
 		toggleUndoButton: jest.fn(),
 		toggleRedoButton: jest.fn(),
+		sendActionButtonPressedAction: jest.fn(),
+		actionButtons: {
+			missingBlockAlertActionButton: 'missing_block_alert_action_button',
+		},
 	};
 } );